Move function Recipe to separate file
